Major Update
Many of the functions completely renewed, and some new functions are added. Many new examples to use, to get familiar with how the system works.
NOTE:  Inertia is not integrated into the MakeLagrange function, and has to be circumvented using RepairInertia. No examples on this yet, though I will try to implement it into the MakeLagrange function some time.
Also NOTE: the format ratio of the animation must be adjusted to the format of you pop-up window, The ones included in the examples are usually good for the full-,half- and small screen sizes.